Add ProShow Screen, partial events screen
All the designs with a certain shade of red were replaced due to different
color being rendered on phone. Screens affected are Events screen and the
bubble on the landing page. The carousel is implemented meaning ProShow
screen and Talks Screen only need the posters to be complete. XML
drawables made for borders to items in events and register screens.